Ontrack MacOS Embedded SSD Imaging Script - 2022
This is an imaging script that will remotely transfer files from one machine to another via rsync and an ssh tunnel via the local wireless network. 
It is meant for device that have no input or outputs available (damaged ports) and will allow for a full filesystem extraction on most Mac OS systems.